Clean up numeric overflow using u128 for product of u64 and 1_000

+++ src/host/jack/stream.rs
@@ -402,9 +402,9 @@ impl jack::ProcessHandler for LocalProcessHandler {
 }
 
 fn micros_to_stream_instant(micros: u64) -> crate::StreamInstant {
-    let nanos = micros * 1000;
+    let nanos = micros as u128 * 1000;
     let secs = micros / 1_000_000;
-    let subsec_nanos = nanos - secs * 1_000_000_000;
+    let subsec_nanos = nanos - secs as u128 * 1_000_000_000;
     crate::StreamInstant::new(secs as i64, subsec_nanos as u32)
 }
